requirements.txt
selenium
fuzzywuzzy
psycopg2
python-dotenv